pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.11"
dependencies = [
    "dj-database-url>=2.3.0",
    "django>=5.1.4",
    "django-bootstrap5>=24.3",
    "django-filter>=24.3",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "pytest-django>=4.9.0",
    "python-dotenv>=1.0.1",
    "rollbar>=1.1.1",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.6",
]
